refactor asset_sync.rake * Ensure consistent Sprockets Environment when `config.assets.compile = false`. * Use Sprockets Environment instead of ActionController asset_path helper to lookup application.js digest path, to ensure consistent behavior in all environments.

+++ dashboard/lib/tasks/asset_sync.rake
@@ -8,22 +8,41 @@ namespace :assets do
   end
 
   task pre_precompile: :environment do
-    next unless CDO.sync_assets
-    require 'cdo/aws/sprockets_cache_s3store'
-    Rails.application.assets.cache = Sprockets::Cache::S3Store.new(
-      CDO.assets_bucket,
-      "#{rack_env}/sprockets_cache.gz"
+    # If rake task is invoked when Rails is configured with `config.assets.compile = false`,
+    # build the Sprockets Environment on `app.assets` explicitly.
+    app = Rails.application
+    app.assets ||= Sprockets::Railtie.build_environment(app)
+
+    if CDO.sync_assets
+      require 'cdo/aws/sprockets_cache_s3store'
+      app.assets.cache = Sprockets::Cache::S3Store.new(
+        CDO.assets_bucket,
+        "#{rack_env}/sprockets_cache.gz"
+      )
+    end
+  end
+
+  task :precompile_application_js do
+    app = Rails.application
+    config = app.config
+    application_js_path = File.join(
+      config.paths['public'].first,
+      config.assets.prefix,
+      app.assets['application.js'].digest_path
     )
+    if File.exist?(application_js_path)
+      puts "minifying", application_js_path
+      uglified = Uglifier.compile(File.read(application_js_path))
+      File.write(application_js_path, uglified)
+    end
   end
 end
 
 Rake::Task['assets:precompile'].enhance([:pre_precompile]) do
-  application_js_path = dashboard_dir('public', ActionController::Base.helpers.asset_path("application.js"))
-  puts "minifying", application_js_path
-  uglified = Uglifier.compile(File.read(application_js_path))
-  File.write(application_js_path, uglified)
-  next unless CDO.sync_assets
-  cache = Rails.application.assets.cache.instance_variable_get(:@cache_wrapper).cache
-  cache.flush if cache.respond_to?(:flush)
-  Rake::Task['assets:sync'].invoke if CDO.sync_assets
+  Rake::Task['assets:precompile_application_js'].invoke
+  if CDO.sync_assets
+    cache = Rails.application.assets.cache.instance_variable_get(:@cache_wrapper).cache
+    cache.flush if cache.respond_to?(:flush)
+    Rake::Task['assets:sync'].invoke
+  end
 end
